day10/requirements.txt
torch>=2.4.0
torchvision>=0.19.0
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.0.0
hydra-core>=1.3.2
tqdm>=4.66.0
PyYAML>=6.0
sam-2
decord>=0.6.0
